windows: 'long' vs. 'int32_t' causes test failure
Switching to the exact type, int32_t (or anything else of
an exact size) vs. 'long' or 'int' etc., which are defined
as a _minimum_ size, allows the test to pass in Windows
and certain other configurations.  For the sake of clarity,
the test is updated to one of the exact types that are
handled in the tests.

+++ tests/common/common/example_messages.h
@@ -302,9 +302,9 @@ struct COMMON_TEST_API
 
 struct COMMON_TEST_API
   MaybeEmpty {
-    static const long DEFAULT_VALUE;
+    static const int32_t DEFAULT_VALUE;
 
-    long value = DEFAULT_VALUE;
+    int32_t value = DEFAULT_VALUE;
 
     RTTR_ENABLE();
   };
